pyproject.toml
[tool.poetry]
name = "fiddle flask"
version = "0.1.0"
description = "Hobby project"
authors = ["Nam Phan <namph.tech@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
Flask = "^2.2.3"
redis = "^4.5.2"
flask-cors = "^3.0.10"
flask-injector = "^0.14.0"
flask-sqlalchemy = "^3.0.3"
flask-caching = "^2.0.2"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
